Extend regex to support original example from issue #820 and more
The original example from issue #820 was using a dash ('-') as fallback,
however, the regular expression did not accept those. It also didn't
support white space (and many other characters) in the fallback, which
are also supported now, so one can even put e.g. links including label
in the fallback.