Allow to hide advanced edit buttons
This only deals with showing or hidding those buttons, but does
not deal with real persmission behind (which are handled by the
back-end).

There is at least one workaround possible for users (apart
from using the JS console): clicking the "pencil" near to the
map name in the edit bar will open the map settings. Not sure
how much we want to block this.

cf #930