Merge OWNER and EDITORS into RESTRICTED level
At Map level, there is no more distinction between owner
and editors, they all share the same rights (but the right
to delete the map, but there is no level to control it).

So having this distinction at DataLayer level does not
make sense anymore, in my opinion.

This is not retrocompatible, as the change at Map level, but
it goes in the sense of permissions management more targetted
to real uMap usage while keeping the system simple.