Prevent computing choropleth breaks for each feature when importing
When importing a choropleth layer, the layer does not have a umap_id,
thus "hasDataLoaded" is immediately true, instead of being true when
all the "known" data is imported.

At this point, I can't find a good reason to keep the check on umap_id,
so let's remove and move forward.