fix dirty flags when re-ordering layers
The index of the top layer in the view starts with zero, while the rank of
the layers count backwards. Thus moving the second-last to the last
position should set the dirty flag of rank 0 and 1.
Instead the former implementation set the dirty flag for layers >= 19 in
a list of 20 layers - resulting in the wrong layers saved.

Fixes #375