Only close ui panel if it is open
When calling ui.closePanel, this will fire "panel:closed", which
is listened by uMap to blindly call map.invalidateSize, which
creates trouble in the dashboard preview: when pressing twice ESC
key (eg. when switching from a preview to another), the map would pan once
to the north east (I guess because invalidateSize is called while
the map is not displayed, but not sure).