fix: deactivate loader on map fragments
Since 8ddc570e23bc403317b0509f50580af9d077ca86 we lost the link
request <=> map, so the "dataloading" and "dataload" events are
global to the DOM, so when loading a list of maps, all instances
react to all dataloading events, which makes the list a mess.

I'd prefer to keep this loading indicator per map, but it's much
easier to just hide it, so let's go that way.