feat: use our custom cluster layer
Overriding MarkerCluster is fragile, and we've made quite a lot of
monkeypatches.
Plus we want to prepare the possibility to cluster semantically, not
only geographicaly. And when we'll have layer groups, to cluster a
whole group (so combining amounts and colors).

This clustering algorithm is very very naive, but uMap is not meant
to be used with millions of points, so it may be enough for our
usages. Let's see and improve later if needed.

Still needed:
- playwright tests
- spiderify function when reaching max zoom and some markers are still
  on the same cluster (eg. same exact position)
- allow to cluster on a property (eg. a region name)
- allow to define a max zoom for clustering (eg. when zooming over the
  "region zoom", let's uncluster)
- allow to display a custom aggregation instead of length in the cluster
  marker (eg. display the total amount of money spent)

fix #326
cf #115
cf #508
cf #2675