fix: moving marker in cluster was not making the map dirty
This was because a race condition between the onDraggend of the
marker (which was then calling a dataChanged, thus a remove/readd
of the marker) and the onDraggend of the marker editor (which is
responsable of setting the marker in dirty state, thus the map).

So this was resulting in the map not being in "savable" state
after moving a marker.