chore: make datalayer dirty when duplicate id has been found
Our first algo to create features id has created a lot of
duplicate, so we need to recompute them, which can take quite
a lot of CPU when there are a lot of features. So let's warn
the editor about this so they can save the layer again, resetting
the ids.