fix: better way to wait for panel transition end
Before this change, we had a corner case when closing
and reopening very fast a panel (example when a full redraw of a
feature is needed, eg. switching from route to normal line),
because the panel was considered closed, but the CSS transition
was never run as the panel was in fact already in its final
position.

Co-authored-by: David Larlet <david@larlet.fr>